UIMA-2252 Added first two views: selection and annotation browser.
Additionally, to the CEV variants, another textfield is added to filter the tree dependent on the covered text of the annotations. Messages are added explaining the usage of the textfields. The icons of the tree are now generated based on the style setting in the CAS Editor: background color indicating the color and two characters indicating the style type (only the first two characters are used). Only all annotations of a type can be shown. When selecting a single annotation in the view, it is selected/focused in the editor.

Both views do already work, but a few small problems remain: The type of the current mode can still be disabled in the tree. Not well tested with different CASs.


git-svn-id: https://svn.apache.org/repos/asf/uima/sandbox/trunk/TextMarker@1185288 13f79535-47bb-0310-9956-ffa450edef68